+++ database/migrations/2021_10_22_082624_add_column_and_relation_user_attribute_table.php
@@ -13,18 +13,19 @@ class AddColumnAndRelationUserAttributeTable extends Migration
      */
     public function up()
     {
-        //
-        Schema::table('user_attribute', function (Blueprint $table) {
-            //
-            $table->integer('sponsor_id')->unsigned()->nullable()->after('user_id');
-        });
+        if(Schema::hasTable('user_attribute')){
+            Schema::table('user_attribute', function (Blueprint $table) {
+                //
+                $table->integer('sponsor_id')->unsigned()->nullable()->after('user_id');
+            });
 
-        Schema::table('user_attribute', function ($table) {
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('restrict')->onUpdate('cascade');
-        });
-        Schema::table('user_attribute', function ($table) {
-            $table->foreign('sponsor_id')->references('sponsor_id')->on('sponsor_master')->onDelete('restrict')->onUpdate('cascade');
-        });
+            Schema::table('user_attribute', function ($table) {
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('restrict')->onUpdate('cascade');
+            });
+            Schema::table('user_attribute', function ($table) {
+                $table->foreign('sponsor_id')->references('sponsor_id')->on('sponsor_master')->onDelete('restrict')->onUpdate('cascade');
+            });
+        }
     }
 
     /**
@@ -34,18 +35,19 @@ class AddColumnAndRelationUserAttributeTable extends Migration
      */
     public function down()
     {
-        //
-        Schema::table('user_attribute', function (Blueprint $table) {
-            $table->dropColumn('sponsor_id');
-        });
-        Schema::table('user_attribute', function (Blueprint $table) {
-            //
-            $table->dropForeign(['user_id']);
-        });
-        Schema::table('user_attribute', function (Blueprint $table) {
-            //
-            $table->dropForeign(['sponsor_id']);
-        });
+        if(Schema::hasTable('user_attribute')){
+            Schema::table('user_attribute', function (Blueprint $table) {
+                $table->dropColumn('sponsor_id');
+            });
+            Schema::table('user_attribute', function (Blueprint $table) {
+                //
+                $table->dropForeign(['user_id']);
+            });
+            Schema::table('user_attribute', function (Blueprint $table) {
+                //
+                $table->dropForeign(['sponsor_id']);
+            });
+        }
         
     }
 }
